+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.2.0](#020)
   - [0.1.2](#012)
   - [0.1.1](#011)
   - [0.1.0](#010)
 
 ---
+
+## 0.2.0
+
+Released on 19/05/2023
+
+- Added Async parser compatible with tokio.
+  - Enable `async` feature (and disable default features if you don't want the sync parser)
 
 ## 0.1.2